Change compilation behavior: console.log() and async-ify
This is similar to how the default devtools console will output the
result of a blob of eval'd javascript. One subtle difference is that a
"return" is required at the end of the CoffeeScript code.

We also keep the IIFE wrapper around the compiled output, and invoke it
asyncronously so that we can freely use await in the CoffeeScript code.